Add JUnit extension to run tests in a ModLauncher context
