+++ src/test/java/org/spongepowered/common/test/event/EventManagerRegistrationTest.java
@@ -25,7 +25,6 @@
 package org.spongepowered.common.test.event;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mockito;
 import org.objectweb.asm.ClassWriter;
 import org.objectweb.asm.MethodVisitor;
@@ -44,14 +43,12 @@ import org.spongepowered.api.event.filter.type.Include;
 import org.spongepowered.api.event.lifecycle.ProvideServiceEvent;
 import org.spongepowered.api.event.lifecycle.StoppedGameEvent;
 import org.spongepowered.api.service.economy.EconomyService;
-import org.spongepowered.common.test.TestEventManager;
-import org.spongepowered.common.test.UnitTestExtension;
+import org.spongepowered.common.launch.event.TestEventManager;
 import org.spongepowered.common.util.DefinableClassLoader;
 import org.spongepowered.plugin.PluginContainer;
 
 import java.lang.reflect.InvocationTargetException;
 
-@ExtendWith(UnitTestExtension.class)
 public class EventManagerRegistrationTest {
 
     @Test
