+++ vanilla/src/applaunch/java/org/spongepowered/vanilla/applaunch/handler/dev/ServerDevLaunchHandler.java
@@ -27,10 +27,10 @@ package org.spongepowered.vanilla.applaunch.handler.dev;
 import cpw.mods.modlauncher.api.ITransformingClassLoader;
 import org.spongepowered.common.applaunch.AppLaunch;
 import org.spongepowered.vanilla.applaunch.AppLaunchTargets;
-import org.spongepowered.vanilla.applaunch.handler.AbstractVanillaLaunchHandler;
+import org.spongepowered.vanilla.applaunch.handler.VanillaLaunchHandler;
 import org.spongepowered.vanilla.applaunch.plugin.VanillaPluginPlatform;
 
-public final class ServerDevLaunchHandler extends AbstractVanillaLaunchHandler {
+public final class ServerDevLaunchHandler extends VanillaLaunchHandler {
 
     @Override
     protected boolean isDev() {
