+++ vanilla/src/launch/java/org/spongepowered/vanilla/launch/VanillaLaunch.java
@@ -31,42 +31,22 @@ import com.google.inject.Module;
 import com.google.inject.Stage;
 import org.spongepowered.common.inject.SpongeCommonModule;
 import org.spongepowered.common.inject.SpongeModule;
-import org.spongepowered.common.launch.Launch;
-import org.spongepowered.common.launch.mapping.SpongeMappingManager;
+import org.spongepowered.common.launch.VanillaBaseLaunch;
 import org.spongepowered.plugin.PluginContainer;
-import org.spongepowered.plugin.metadata.PluginMetadata;
-import org.spongepowered.plugin.metadata.builtin.MetadataContainer;
-import org.spongepowered.plugin.metadata.builtin.MetadataParser;
 import org.spongepowered.vanilla.applaunch.plugin.VanillaPluginPlatform;
 import org.spongepowered.vanilla.launch.inject.SpongeVanillaModule;
-import org.spongepowered.vanilla.launch.mapping.VanillaMappingManager;
-import org.spongepowered.vanilla.launch.plugin.VanillaDummyPluginContainer;
 import org.spongepowered.vanilla.launch.plugin.VanillaPluginManager;
 
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.Reader;
-import java.net.URISyntaxException;
-import java.net.URL;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.Enumeration;
 import java.util.List;
 
-public abstract class VanillaLaunch extends Launch {
+public abstract class VanillaLaunch extends VanillaBaseLaunch {
 
     private final Stage injectionStage;
-    private final VanillaPluginManager pluginManager;
-    private final VanillaMappingManager mappingManager;
     private PluginContainer vanillaPlugin;
 
     protected VanillaLaunch(final VanillaPluginPlatform pluginPlatform, final Stage injectionStage) {
-        super(pluginPlatform);
+        super(pluginPlatform, new VanillaPluginManager());
         this.injectionStage = injectionStage;
-        this.pluginManager = new VanillaPluginManager();
-        this.mappingManager = new VanillaMappingManager();
     }
 
     @Override
@@ -93,16 +73,6 @@ public abstract class VanillaLaunch extends Launch {
     }
 
     @Override
-    public final VanillaPluginManager pluginManager() {
-        return this.pluginManager;
-    }
-
-    @Override
-    public SpongeMappingManager mappingManager() {
-        return this.mappingManager;
-    }
-
-    @Override
     public Injector createInjector() {
         final List<Module> modules = Lists.newArrayList(
             new SpongeModule(),
@@ -110,69 +80,5 @@ public abstract class VanillaLaunch extends Launch {
             new SpongeVanillaModule()
         );
         return Guice.createInjector(this.injectionStage(), modules);
-    }
-
-    protected final void launchPlatform(final String[] args) {
-        this.createPlatformPlugins();
-        this.logger().info("Loading Sponge, please wait...");
-        this.performBootstrap(args);
-    }
-
-    protected abstract void performBootstrap(final String[] args);
-
-    protected final void createPlatformPlugins() {
-        final String metadataFileLocation = this.pluginPlatform.metadataFilePath();
-        try {
-            // This is a bit nasty, but allows Sponge to detect builtin platform plugins when it's not the first entry on the classpath.
-            final URL classUrl = VanillaLaunch.class.getResource("/" + VanillaLaunch.class.getName().replace('.', '/') + ".class");
-
-            MetadataContainer read = null;
-
-            // In production, let's try to ensure we can find our descriptor even if we're not first on the classpath
-            if (classUrl.getProtocol().equals("jar")) {
-                // Extract the path of the underlying jar file, and parse it as a path to normalize it
-                final String[] classUrlSplit = classUrl.getPath().split("!");
-                final Path expectedFile = Paths.get(new URL(classUrlSplit[0]).toURI());
-
-                // Then go through every possible resource
-                final Enumeration<URL> manifests =
-                        VanillaLaunch.class.getClassLoader().getResources("/" + metadataFileLocation);
-                while (manifests.hasMoreElements()) {
-                    final URL next = manifests.nextElement();
-                    if (!next.getProtocol().equals("jar")) {
-                        continue;
-                    }
-
-                    // And stop when the normalized jar in that resource matches the URL of the jar that loaded VanillaLaunch?
-                    final String[] pathSplit = next.getPath().split("!");
-                    if (pathSplit.length == 2) {
-                        final Path vanillaPath = Paths.get(new URL(pathSplit[0]).toURI());
-                        if (vanillaPath.equals(expectedFile)) {
-                            try (final Reader reader = new InputStreamReader(next.openStream(), StandardCharsets.UTF_8)) {
-                                read = MetadataParser.read(reader);
-                            }
-                            break;
-                        }
-                    }
-                }
-            }
-
-            if (read == null) { // other measures failed, fall back to directly querying the classpath
-                final Path vanillaPath =
-                        Paths.get(VanillaLaunch.class.getResource("/" + metadataFileLocation).toURI());
-                try (final Reader reader = Files.newBufferedReader(vanillaPath, StandardCharsets.UTF_8)) {
-                    read = MetadataParser.read(reader);
-                }
-            }
-            if (read == null) {
-                throw new RuntimeException("Could not determine location for implementation metadata!");
-            }
-
-            for (final PluginMetadata metadata : read.metadata()) {
-                this.pluginManager().addPlugin(new VanillaDummyPluginContainer(metadata, this.logger(), this));
-            }
-        } catch (final IOException | URISyntaxException e) {
-            throw new RuntimeException("Could not load metadata information for the implementation! This should be impossible!");
-        }
     }
 }
